requirements.txt
requests
discord
python-dotenv